Create streamlit-model.py
The purpose of this project was to develop a capstone of our choice. 

For this, I decided to use Streamlit to build a simple linear regression model between two quantitative samples and determine its scatter plot, line of best fit, intercept, slope, Pearson coefficient, coefficient of determination (r-square value) and regression model. 

From the regression model, make predictions of the response variable given the explanatory variable.